+++ CHANGELOG.md
@@ -1,7 +1,5 @@
 # Changelog
 
-## Info arch
-
 - Add `Breadcrumb` component. This component is included as a feature in `PageLayout`. [#331](https://github.com/mapbox/dr-ui/pull/331).
 - Add `OnThisPage` component. This component displays headings on the page and performs scroll spy to indicate where you are on the page.
   - 🚨This update removes the need for the `@mapbox/dr-ui/plugins/create-sections` plugin and it is now deprecated. You can safely remove it from your `batfish.config.js`.
@@ -31,6 +29,7 @@
 - Allow passing children to the `ExamplesIndex` component.
 - Add theming options to `OverviewHeader`: `theme` to accept CSS classes to customize the container, `lightText` to enable white text, `description` to add a text description. The `features` prop is now optional.
 - 🚨Deprecate `GlossaryCard`, `GlossaryPage`, and `GlossarySection` components.
+- Improve color contrast on `LevelIndicator`.
 
 ## 1.3.0